Finished initial refactoring
